Various fixes for Linux and recent gcc releases (missing headers, incorrect conversions). Now using C++ version of all C headers (such as <math.h> to <cmath>). Added Unicode Linux versions of the Utils lib. Added an output to the console when running debug binaries.

+++ trunk/src/gui/frost_text.cpp
@@ -485,7 +485,7 @@ namespace Frost
             foreach (iterLine, lLines)
             {
                 lLineList_.push_back(*iterLine);
-                if (s_uint(lLineList_.size()) == uiMaxLineNbr)
+                if (s_uint((uint)lLineList_.size()) == uiMaxLineNbr)
                 {
                     return;
                 }
